--- a/daml-lf/interpreter/perf/src/main/scala/com/daml/lf/explore/Explore.scala
+++ b/daml-lf/interpreter/perf/src/main/scala/com/daml/lf/explore/Explore.scala
@@ -30,8 +30,8 @@ object PlaySpeedy {
 
     names.foreach { name =>
       val (expected, expr) = examples(name)
-      val converted = compiler.unsafeClosureConvert(expr)
-      val machine = Speedy.Machine.fromPureSExpr(noPackages, converted)
+      val anf = compiler.unsafeCompilationPipeline(expr)
+      val machine = Speedy.Machine.fromPureAExpr(noPackages, anf)
       runMachine(name, machine, expected)
     }
   }
@@ -84,7 +84,7 @@ object PlaySpeedy {
     }
   }
 
-  final case class MachineProblem(s: String) extends RuntimeException(s, null, false, false)
+  final case class MachineProblem(s: String) extends RuntimeException(s)
 
   def examples: Map[String, (Int, SExpr)] = {
 
@@ -97,6 +97,8 @@ object PlaySpeedy {
 
     def subtract2(x: SExpr, y: SExpr): SExpr = SEApp(SEBuiltin(SBSubInt64), Array(x, y))
     val subtract = SEAbs(2, subtract2(SEVar(2), SEVar(1)))
+
+    val identity = SEAbs(1, SEVar(1))
 
     def twice2(f: SExpr, x: SExpr): SExpr = SEApp(f, Array(SEApp(f, Array(x))))
     val twice = SEAbs(2, twice2(SEVar(2), SEVar(1)))
@@ -117,6 +119,10 @@ object PlaySpeedy {
         "subF", //88-55
         33,
         SEApp(subtract, Array(num(88), num(55)))),
+      (
+        "id-id", // id id 77
+        77,
+        SEApp(identity, Array(identity, num(77)))),
       (
         "thrice", // thrice (\x -> x - 1) 0
         -3,