… compiler/damlc/tests/daml-test-files/OverAppliedFoldEvalOrder.daml
… compiler/damlc/tests/daml-test-files/SemanticsEvalOrder.daml
… daml-lf/interpreter/src/main/scala/com/digitalasset/daml/lf/speedy/Classify.scala
… daml-lf/interpreter/src/main/scala/com/digitalasset/daml/lf/speedy/Compiler.scala
… daml-lf/interpreter/src/main/scala/com/digitalasset/daml/lf/speedy/Pretty.scala
… daml-lf/interpreter/src/main/scala/com/digitalasset/daml/lf/speedy/PrettyLightweight.scala
… daml-lf/interpreter/src/main/scala/com/digitalasset/daml/lf/speedy/SBuiltin.scala
… daml-lf/interpreter/src/main/scala/com/digitalasset/daml/lf/speedy/SExpr.scala
… daml-lf/interpreter/src/main/scala/com/digitalasset/daml/lf/speedy/SValue.scala
… daml-lf/interpreter/src/main/scala/com/digitalasset/daml/lf/speedy/Speedy.scala
… daml-lf/interpreter/src/test/scala/com/digitalasset/daml/lf/speedy/AnfTest.scala
… daml-lf/repl/src/main/scala/com/digitalasset/daml/lf/repl/testing/Main.scala
… daml-lf/scenario-interpreter/src/perf/benches/scala/com/digitalasset/daml/lf/speedy/perf/CollectAuthority.scala
… daml-script/runner/src/main/scala/com/digitalasset/daml/lf/engine/script/Runner.scala
… daml-script/test/src/com/digitalasset/daml/lf/engine/script/test/SingleParticipant.scala
… triggers/runner/src/main/scala/com/digitalasset/daml/lf/engine/trigger/Runner.scala