Safe ANF transformation for Speedy.

This PR reinstates the ANF transformation as part of the Speedy compilation phase. The original PR #6440 changed evaluation order in some pathological cases, and so was reverted by PR #6645.

This new PR modifies the original work to be careful not to apply the ANF transform in cases where it can't be sure the evaluation order wont change. This PR was prepared by reinstating the reverting change & working from there. Specifically, we have made the following steps:

- (1) `9665562af` reinstate ANF transform; get FAIL for EvaluationOrder test
  
  This undoes the effect of the revert commit but keeps the `EvaluationOrder` test which shows the bug in full ANF.

      git revert 02c59d4f2
      git reset HEAD^
      git checkout compiler/damlc/tests/daml-test-files/EvaluationOrder.daml

- (2) `f66047fc8` Bend `scalafmt` to my will.

  This change cherry picks commit `992b1dc1d` from PR: #6641 which was already approved, but never got merged before ANF was reverted.  We bring it in here because it makes `Anf.scala` vastly more readable.

- (3) `41be460e7` fix ANF transform to be safe for evaluation order

  The ANF transform is applied only when it is safe to do so. We defined safe cases as: (a) an application with only a single argument, and (b) an application when the expression in function position is a builtin.

  There is some potential for more safe cases, but the returns are diminishing since most applications either fall in the above two categories, or else are cases we have no hope of ever proving as safe.

  In the unsafe cases we expand the multi-application into a cascade of simple applications, and perform ANF on the result. This does not change the evaluation order!

- (4) `777b4962d` revert location change in `traceMsg`. The location is unchanged by safe ANF

  The modified ANF transform doesn't change the location. Which is good I guess.

- (5) `aa450e705` adapt AnfTests to show modified (safe) ANF transform

  Update tests to show the modified transform where necessary.


### Performance

Here is the perf for `CollectAuthority`, running interleaved 6-way vs master on my cloud machine. About x1.15 speedup.

    here:    55.141 ±(99.9%) 0.477 ms/op [Average]
    master:  63.887 ±(99.9%) 0.336 ms/op [Average]
    here:    56.083 ±(99.9%) 0.511 ms/op [Average]
    master:  62.538 ±(99.9%) 0.351 ms/op [Average]
    here:    54.814 ±(99.9%) 0.234 ms/op [Average]
    master:  63.585 ±(99.9%) 0.198 ms/op [Average]


### Pull Request Checklist

- [ ] Read and understand the [contribution guidelines](https://github.com/digital-asset/daml/blob/master/CONTRIBUTING.md)
- [ ] Include appropriate tests
- [ ] Set a descriptive title and thorough description
- [ ] Add a reference to the [issue this PR will solve](https://github.com/digital-asset/daml/issues), if appropriate
- [ ] Include changelog additions in one or more commit message bodies between the `CHANGELOG_BEGIN` and `CHANGELOG_END` tags
- [ ] Normal production system change, include purpose of change in description

NOTE: CI is not automatically run on non-members pull-requests for security
reasons. The reviewer will have to comment with `/AzurePipelines run` to
trigger the build.