@nickchapman-da 2020-07-20
rebased on master, and fixed conflict