@cocreature 2020-07-22
Fantastic work, thank you very much!
I’ll leave the detailed review to @hurryabit but I got a bit confused by your fix for recursive builtins and spent some time thinking about it. It looks like this is unfortunately not sufficient to fix the issue. If you look at your example, there are two things that are crucial to cause the error (before your fix):

1. A fully applied call to `foldr` can crash. This holds true for a lot of builtins (error, division by zero, basically anything around the generic comparisons, …).
2. It can return a function which means it can be overapplied while still passing the typechecker. This applies to only very few functions. From a quick look, the only non-recursive builtin is `error`.

Note that neither of those properties are recursion but I wouldn’t be surprised if recursion does play a role that I don’t understand.

Based on those two points, you can construct an example using `error` that still changes behavior here.

```
test = (error "abc") (error "def")
```
(I’m assuming that this is DAML-LF and directly translates to primitives. If you wrap it in non-primitives then your optimizations for non-recursive primitives don’t kick in).

I’ve pushed an executable that produces a DAR with this to https://github.com/digital-asset/daml/tree/overapplied-builtin. On master it errors with `abc`. On this branch it errors with `def`.

I think the right solution here is to never overapply builtins. We have the arity of builtins so we can split at that point. For most primitives that does not change anything to the current state since they cannot be overapplied so this does not change anything. For primitives that can be overapplied it fixes the issue both for recursive and non-recursive builtins and for recursive builtins it would actually be better than what you currently implemented since you can apply them to all (3) arguments at once instead of splitting it into unary applications.