@nickchapman-da 2020-07-22
Thanks for your review @cocreature . And good catch on the `error` builtin. I was under the mistaken impression that no builtin could return an function, and so be over-applied. But clearly this can happen for error.

And it makes sense to check all builtin applications against their known arity, and only regard application of up-to this number as safe. Then we are somewhat future proof against future builtins.