@nickchapman-da 2020-07-22
> Does this testcase fail wtihout the fix? I tried this earlier and couldn’t get it to fail since `error` is not translated directly into the builtin. That’s why my testcase contains handwritten LF.

No it doesn't. Which is very sad.
So that's why I added to the AnfTests.scala