@nickchapman-da 2020-07-23
> I reviewed under the assumption that this is basically the old ANF PR + changes to how we transform applications in order to maintain safety. Is this assumption justified?

Yes.