@nickchapman-da 2020-07-29
This PR has been superseded by https://github.com/digital-asset/daml/pull/6918